% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customers.R
\name{customers.update}
\alias{customers.update}
\title{Updates a customer.}
\usage{
customers.update(
  customerKey,
  kind = NULL,
  customerDomain = NULL,
  alternateEmail = NULL,
  phoneNumber = NULL,
  etag = NULL,
  id = NULL,
  customerCreationTime = NULL,
  postalAddress = NULL,
  language = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["directory.customers.update"]]$scopes),
  return_response = F
)
}
\arguments{
\item{customerKey}{Id of the customer to be updated}

\item{kind}{Identifies the resource as a customer. Value: \code{admin#directory#customer}}

\item{customerDomain}{The customer's primary domain name string. Do not include the \code{www} prefix when creating a new customer.}

\item{alternateEmail}{The customer's secondary contact email address. This email address cannot be on the same domain as the \code{customerDomain}}

\item{phoneNumber}{The customer's contact phone number in \href{https://en.wikipedia.org/wiki/E.164}{E.164} format.}

\item{etag}{ETag of the resource.}

\item{id}{The unique ID for the customer's Google Workspace account. (Readonly)}

\item{customerCreationTime}{The customer's creation time (Readonly)}

\item{postalAddress}{The customer's postal address information.}

\item{language}{The customer's ISO 639-2 language code. See the \href{/admin-sdk/directory/v1/languages}{Language Codes} page for the list of supported codes. Valid language codes outside the supported set will be accepted by the API but may lead to unexpected behavior. The default value is \code{en}.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{package_make}}.
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/admin.directory.customer
}
}
\seealso{
\href{http://developers.google.com/admin-sdk/}{Google Documentation}
}
