% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{groups.list}
\alias{groups.list}
\title{List Groups}
\usage{
groups.list(
  customer = NULL,
  query = NULL,
  sortOrder = NULL,
  pageToken = NULL,
  orderBy = NULL,
  userKey = NULL,
  domain = NULL,
  maxResults = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["directory.groups.list"]]$scopes),
  return_response = F
)
}
\arguments{
\item{customer}{The unique ID for the customer's Google Workspace account. In case of a multi-domain account, to fetch all groups for a customer, fill this field instead of domain. As an account administrator, you can also use the \code{my_customer} alias to represent your account's \code{customerId}. The \code{customerId} is also returned as part of the \href{/admin-sdk/directory/v1/reference/users}{Users}}

\item{query}{Query string search. Should be of the form "". Complete documentation is at https: //developers.google.com/admin-sdk/directory/v1/guides/search-groups}

\item{sortOrder}{Whether to return results in ascending or descending order. Only of use when orderBy is also used}

\item{pageToken}{Token to specify next page in the list}

\item{orderBy}{Column to use for sorting results}

\item{userKey}{Email or immutable ID of the user if only those groups are to be listed, the given user is a member of. If it's an ID, it should match with the ID of the user object.}

\item{domain}{The domain name. Use this field to get fields from only one domain. To return all domains for a customer account, use the \code{customer} query parameter instead.}

\item{maxResults}{Maximum number of results to return. Max allowed value is 200.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{package_make}}.
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/admin.directory.group
\item https://www.googleapis.com/auth/admin.directory.group.readonly
}
}
\seealso{
\href{http://developers.google.com/admin-sdk/}{Google Documentation}
}
