% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{groups.patch}
\alias{groups.patch}
\title{Updates a group's properties.}
\usage{
groups.patch(
  groupKey,
  adminCreated = NULL,
  id = NULL,
  name = NULL,
  description = NULL,
  etag = NULL,
  nonEditableAliases = NULL,
  directMembersCount = NULL,
  kind = NULL,
  aliases = NULL,
  email = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["directory.groups.patch"]]$scopes),
  return_response = F
)
}
\arguments{
\item{groupKey}{Identifies the group in the API request. The value can be the group's email address, group alias, or the unique group ID.}

\item{adminCreated}{Value is \code{true} if this group was created by an administrator rather than a user.}

\item{id}{The unique ID of a group. A group \code{id} can be used as a group request URI's \code{groupKey}.}

\item{name}{The group's display name.}

\item{description}{An extended description to help users determine the purpose of a group. For example, you can include information about who should join the group, the types of messages to send to the group, links to FAQs about the group, or related groups. Maximum length is \verb{4,096} characters.}

\item{etag}{ETag of the resource.}

\item{nonEditableAliases}{List of the group's non-editable alias email addresses that are outside of the account's primary domain or subdomains. These are functioning email addresses used by the group. This is a read-only property returned in the API's response for a group. If edited in a group's POST or PUT request, the edit is ignored by the API service.}

\item{directMembersCount}{The number of users that are direct members of the group. If a group is a member (child) of this group (the parent), members of the child group are not counted in the \code{directMembersCount} property of the parent group.}

\item{kind}{The type of the API resource. For Groups resources, the value is \code{admin#directory#group}.}

\item{aliases}{List of a group's alias email addresses.}

\item{email}{The group's email address. If your account has multiple domains, select the appropriate domain for the email address. The \code{email} must be unique. This property is required when creating a group. Group email addresses are subject to the same character usage rules as usernames, see the \href{https://support.google.com/a/answer/9193374}{help center} for details.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
This method supports \href{/admin-sdk/directory/v1/guides/performance#patch}{patch semantics}. Autogenerated via \code{\link[googlePackageMaker]{package_make}}.
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/admin.directory.group
}
}
\seealso{
\href{http://developers.google.com/admin-sdk/}{Google Documentation}
}
