% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/members.R
\name{members.patch}
\alias{members.patch}
\title{Updates the membership properties of a user in the specified group.}
\usage{
members.patch(
  memberKey,
  groupKey,
  type = NULL,
  kind = NULL,
  role = NULL,
  etag = NULL,
  email = NULL,
  status = NULL,
  id = NULL,
  delivery_settings = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["directory.members.patch"]]$scopes),
  return_response = F
)
}
\arguments{
\item{memberKey}{Identifies the group member in the API request. A group member can be a user or another group. The value can be the member's (group or user) primary email address, alias, or unique ID.}

\item{groupKey}{Identifies the group in the API request. The value can be the group's email address, group alias, or the unique group ID.}

\item{type}{The type of group member.}

\item{kind}{The type of the API resource. For Members resources, the value is \code{admin#directory#member}.}

\item{role}{The member's role in a group. The API returns an error for cycles in group memberships. For example, if \code{group1} is a member of \code{group2}, \code{group2} cannot be a member of \code{group1}. For more information about a member's role, see the \href{https://support.google.com/a/answer/167094}{administration help center}.}

\item{etag}{ETag of the resource.}

\item{email}{The member's email address. A member can be a user or another group. This property is required when adding a member to a group. The \code{email} must be unique and cannot be an alias of another group. If the email address is changed, the API automatically reflects the email address changes.}

\item{status}{Status of member (Immutable)}

\item{id}{The unique ID of the group member. A member \code{id} can be used as a member request URI's \code{memberKey}.}

\item{delivery_settings}{Defines mail delivery preferences of member. This is only supported by create/update/get.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
This method supports \href{/admin-sdk/directory/v1/guides/performance#patch}{patch semantics}. Autogenerated via \code{\link[googlePackageMaker]{package_make}}.
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/admin.directory.group
\item https://www.googleapis.com/auth/admin.directory.group.member
}
}
\seealso{
\href{http://developers.google.com/admin-sdk/}{Google Documentation}
}
