% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas.R
\name{schemas.get}
\alias{schemas.get}
\title{NA}
\usage{
schemas.get(
  schemaKey,
  customerId,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["directory.schemas.get"]]$scopes),
  return_response = F
)
}
\arguments{
\item{schemaKey}{Name or immutable ID of the schema.}

\item{customerId}{Immutable ID of the Google Workspace account.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
NA Autogenerated via \code{\link[googlePackageMaker]{package_make}}.
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/admin.directory.userschema
\item https://www.googleapis.com/auth/admin.directory.userschema.readonly
}
}
\seealso{
\href{http://developers.google.com/admin-sdk/}{Google Documentation}
}
